fix: enable DuckDB Docker builds
